+++ SignNow.Net.Examples/ExamplesRunner.cs
@@ -15,8 +15,10 @@ using SignNow.Net.Model;
 using SignNow.Net.Model.ComplexTags;
 using SignNow.Net.Model.EditFields;
 using SignNow.Net.Model.Requests;
+using SignNow.Net.Model.Requests.EventSubscriptionBase;
 using SignNow.Net.Model.Requests.GetFolderQuery;
 using SignNow.Net.Test.Context;
+using UnitTests;
 
 namespace SignNow.Net.Examples
 {
@@ -850,6 +852,73 @@ namespace SignNow.Net.Examples
                 .ConfigureAwait(false);
 
             Assert.IsFalse(folders.Folders.Any(f => f.Name == "DeleteMe"));
+        }
+
+        #endregion
+
+        #region WebHooks Examples
+
+        /// <summary>
+        /// Allows to subscribe an external service(callback_url) to a specific event of user or document.
+        /// As soon as a certain selected event from the List of event types occurs, SignNow sends a notification about it.
+        /// </summary>
+        /// <see cref="https://docs.signnow.com/docs/signnow/branches/v1.2/reference/operations/create-a-api-v-2-event"/>
+        [TestMethod]
+        public async Task CreateEventSubscription()
+        {
+            // Upload document with fields
+            await using var fileStream = File.OpenRead(PdfWithSignatureField);
+            var document = await testContext.Documents
+                .UploadDocumentWithFieldExtractAsync(fileStream, "DocumentForEventSubscriptionCreate.pdf")
+                .ConfigureAwait(false);
+
+            // Using signNowContext lets create event subscription
+            var myCallbackUrl = new Uri("https://signnow.com/callbackHandler");
+            await testContext.Events
+                .CreateEventSubscriptionAsync(
+                    new CreateEventSubscription(EventType.DocumentComplete, document.Id, myCallbackUrl))
+                .ConfigureAwait(false);
+
+            // Check for successful created event subscription
+            // Gets information about all subscriptions to events made with a specific application
+            // <see cref="https://docs.signnow.com/docs/signnow/reference/operations/list-api-v-2-events"/>
+            var eventSubscriptionList = await testContext.Events
+                .GetEventSubscriptionsAsync(new PagePaginationOptions { Page = 1, PerPage = 1})
+                .ConfigureAwait(false);
+
+            // Determining events total to get the latest event from whole events list
+            var latestPage = eventSubscriptionList.Meta.Pagination.Total;
+
+            // Getting event details from Events list
+            var myLatestCreatedEvent = await testContext.Events
+                .GetEventSubscriptionsAsync(new PagePaginationOptions { Page = latestPage, PerPage = 1})
+                .ConfigureAwait(false);
+
+            var myLatestEvent = myLatestCreatedEvent.Data.First();
+            Assert.AreEqual(EventType.DocumentComplete, myLatestEvent.Event);
+            Assert.AreEqual(myCallbackUrl, myLatestEvent.JsonAttributes.CallbackUrl);
+
+            // Changing an existing event subscription
+            // <see cref="https://docs.signnow.com/docs/signnow/reference/operations/update-a-api-v-2-event"/>
+            var eventForUpdate = myLatestCreatedEvent.Data.First();
+            eventForUpdate.JsonAttributes.CallbackUrl = new Uri("https://signnow.com/myNewCallbackHandler");
+
+            var changedEvent = await testContext.Events
+                .UpdateEventSubscriptionAsync(new UpdateEventSubscription(eventForUpdate))
+                .ConfigureAwait(false);
+
+            var updatedEvent = await testContext.Events
+                .GetEventSubscriptionInfoAsync(changedEvent.Id)
+                .ConfigureAwait(false);
+
+            Assert.AreEqual(myLatestEvent.Id, updatedEvent.Id);
+            Assert.AreEqual("https://signnow.com/myNewCallbackHandler", updatedEvent.JsonAttributes.CallbackUrl.AbsoluteUri);
+
+            // Unsubscribes an external service (callback_url) from specific events of user or document
+            // <see cref="https://docs.signnow.com/docs/signnow/reference/operations/delete-a-api-v-2-event"/>
+            await testContext.Events
+                .DeleteEventSubscriptionAsync(myLatestEvent.Id)
+                .ConfigureAwait(false);
         }
 
         #endregion