+++ SignNow.Net.Test/UnitTests/Models/RequestOptionsTest.cs
@@ -2,6 +2,7 @@ using System;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using SignNow.Net.Internal.Requests;
 using SignNow.Net.Model;
+using SignNow.Net.Model.Requests;
 using Method = System.Net.Http.HttpMethod;
 
 namespace UnitTests.Models
@@ -16,7 +17,7 @@ namespace UnitTests.Models
         [TestInitialize]
         public void Setup()
         {
-            content = new JsonHttpContent(new { document_id = "test" });
+            content = new CreateSigningLinkRequest { DocumentId = "test" };
             requestUrl = new Uri("https://signnow.com");
             token = new Token
             {