+++ SignNow.Net/_Internal/Requests/CreateSigningLinkRequest.cs
@@ -1,0 +1,11 @@
+using Newtonsoft.Json;
+using SignNow.Net.Model.Requests;
+
+namespace SignNow.Net.Internal.Requests
+{
+    internal class CreateSigningLinkRequest : JsonHttpContent
+    {
+        [JsonProperty("document_id")]
+        public string DocumentId { get; set; }
+    }
+}