+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 and this project adheres to [Semantic Versioning](http://semver.org).
 
 ## [Unreleased] - TBD
+### Added
+- `IEventSubscription` interface for any operation with Events
+- Webhooks 2.0 API are implemented
+
+### Changed
+- Token type is always set to `Bearer` or `Basic` explicitly in every service
 
 
 ## [1.2.1] - 2023-07-13